made it work on rhel

+++ src/main/java/gov/fema/geo/LayerRepository.java
@@ -2,8 +2,10 @@ package gov.fema.geo;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface LayerRepository extends CrudRepository<Layer, Integer> {
+public interface LayerRepository extends CrudRepository<Layer, Long> {
 
-    Layer findLayerById(Integer id);
+    Layer findLayerById(Long id);
+    //https://www.baeldung.com/spring-data-crud-repository-save#updateInstance
+    // not needed? void updateLayer(Layer layer);
 }
